Move definitions from main_test.go to utility.go
runCommand and pickAvailablePort functions are not visible in other
packages during build because the are defined in _test.go package.
Moving them into utility.go and thus making them part of the package.
