+++ resources/test/main_test.go
@@ -4,11 +4,8 @@ import (
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
-	"net"
 	"os"
-	"os/exec"
 	"path/filepath"
-	"strings"
 	"syscall"
 	"testing"
 	"time"
@@ -24,14 +21,6 @@ type mnemonicFile struct {
 
 type agentFile struct {
 	Agent string
-}
-
-type jobFile struct {
-	Job string
-}
-
-type jobInvocationFile struct {
-	Signature string
 }
 
 var testConfiguration = []string{
@@ -123,33 +112,4 @@ func TestEndToEnd(t *testing.T) {
 	header("Testing gRPC-web client")
 	testGRPCWeb(t, h, daemonPort)
 	time.Sleep(time.Second)
-}
-
-func runCommand(dir string, env []string, name string, arg ...string) *exec.Cmd {
-	cmd := exec.Command(name, arg...)
-	if dir != "" {
-		cmd.Dir = dir
-	}
-	if env != nil && len(env) > 0 {
-		cmd.Env = append(os.Environ(), env...)
-	}
-	cmd.Stderr = os.Stderr
-	cmd.Stdout = os.Stdout
-	cmd.Start()
-	return cmd
-}
-
-func pickAvailablePort() string {
-	p, err := net.Listen("tcp", ":0")
-	if err != nil {
-		panic(err)
-	}
-	addr := p.Addr().String()
-	parts := strings.Split(addr, ":")
-	if len(parts) < 2 {
-		panic("Can't parse address: " + addr)
-	}
-	p.Close()
-
-	return parts[len(parts)-1]
 }
